+++ packages/docs/.vitepress/theme/components/Demo.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="my-4">
-    <div class="flex justify-start aspect-2/1">
+    <div class="flex flex-col aspect-2/1">
       <component
         :is="componentName"
       />