+++ packages/react/src/hooks/useLatest.ts
@@ -1,10 +1,10 @@
 'use client'
 
-import type { MutableRefObject } from 'react'
+import type { RefObject } from 'react'
 
 import { useRef } from 'react'
 
-export function useLatest<T>(value: T): MutableRefObject<T> {
+export function useLatest<T>(value: T): RefObject<T> {
   const ref = useRef(value)
   ref.current = value
   return ref